+++ .circleci/config.yml
@@ -10,7 +10,7 @@ jobs:
       - checkout
       - restore_cache:
           key: &java_cache_key v2-{{ checksum "build.gradle.kts" }}
-      - run: 
+      - run:
           command: "bash gradlew check"
       - save_cache:
           paths:
@@ -18,26 +18,17 @@ jobs:
           key: *java_cache_key
       - store_test_results:
           path: ~/cookbook/build/reports/allTestsXml
-  # build_ui:
-  #   docker:
-  #     - image: circleci/openjdk:8u181-jdk-node-browsers
-  #   steps:
-  #     - checkout
-  #     - run:
-  #         working_directory: ~/project/punchcard-ui
-  #         command: yarn
-  #     - run: 
-  #         working_directory: ~/project/punchcard-ui
-  #         command: yarn test:ci
-  #     - run: 
-  #         working_directory: ~/project/punchcard-ui
-  #         command: mkdir jest & mv junit.xml jest
-  #     - store_test_results:
-  #         path: ~/project/punchcard-ui/jest
+  build_ui:
+    docker:
+      - image: circleci/openjdk:8u181-jdk-node-browsers
+    steps:
+      - checkout
+      - run: yarn workspaces run test
+      - run: yarn workspaces run build
 
 workflows:
   version: 2
   full_build:
     jobs:
     - build_java
-    # - build_ui
+    - build_ui
