+++ BUILD
@@ -1,3 +1,1 @@
-package(default_visibility = ["//visibility:public"])
 
-exports_files(["tsconfig.json"])
