--- lerna.json
@@ -1,8 +1,0 @@
-{
-  "packages": [
-    "packages/*"
-  ],
-  "version": "0.1.0",
-  "npmClient": "yarn",
-  "useWorkspaces": true
-}
