+++ package.json
@@ -9,6 +9,9 @@
     "packages/*"
   ],
   "devDependencies": {
-    "lerna": "^3.13.1"
+    "json-server": "^0.14.2"
+  },
+  "scripts": {
+    "punchcard:backend": "json-server punchcard-db.json --port=8080"
   }
 }
