+++ packages/punchcard-ui/package.json
@@ -7,8 +7,12 @@
     "@types/node": "11.11.1",
     "@types/react": "16.8.7",
     "@types/react-dom": "16.8.2",
+    "@types/react-router-dom": "^4.3.1",
+    "axios": "^0.18.0",
     "react": "^16.8.4",
     "react-dom": "^16.8.4",
+    "react-heatmap-grid": "^0.7.3",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.8",
     "typescript": "3.3.3333"
   },
@@ -27,25 +31,5 @@
     "not ie <= 11",
     "not op_mini all"
   ],
-  "description": "UI for Strava Punchcard application",
-  "author": "Peter Teixeira <PtrTeixeira@gmail.com>",
-  "homepage": "http://github.com/PtrTeixeira/cookbook",
-  "license": "MIT",
-  "directories": {
-    "lib": "src",
-    "test": "test"
-  },
-  "files": [
-    "src"
-  ],
-  "publishConfig": {
-    "registry": "https://registry.yarnpkg.com"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/PtrTeixeira/cookbook.git"
-  },
-  "bugs": {
-    "url": "https://github.com/PtrTeixeira/cookbook/issues"
-  }
+  "description": "UI for Strava Punchcard application"
 }
