+++ packages/punchcard-ui/src/@types/react-heatmap-grid/index.d.ts
@@ -1,0 +1,10 @@
+declare module 'react-heatmap-grid' {
+  export interface HeatmapProps {
+    xLabels: string[];
+    yLabels: string[];
+    data: number[][];
+  }
+
+  declare const HeatMap: React.SFC<HeatmapProps>;
+  export default HeatMap;
+}
