+++ packages/punchcard-ui/src/App.test.tsx
@@ -1,8 +1,13 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+
+import 'jest-localstorage-mock';
+
 import App from './App';
 
 it('renders without crashing', () => {
+  sessionStorage.setItem('results', "{}")
+
   const div = document.createElement('div');
   ReactDOM.render(<App />, div);
   ReactDOM.unmountComponentAtNode(div);
