+++ punchcard-db.json
@@ -1,0 +1,32 @@
+{
+    "punchcard": {
+        "MONDAY": {
+            "18": 4,
+            "19": 1
+        },
+        "TUESDAY": {
+            "18": 4,
+            "19": 1
+        },
+        "WEDNESDAY": {
+            "18": 4,
+            "19": 1
+        },
+        "THURSDAY": {
+            "18": 4,
+            "19": 1
+        },
+        "FRIDAY": {
+            "18": 4,
+            "19": 1
+        },
+        "SATURDAY": {
+            "13": 6,
+            "14": 4
+        },
+        "SUNDAY": {
+            "8": 4,
+            "9": 10
+        }
+    }
+}
