--- punchcard/ui/.gitignore
@@ -1,3 +1,0 @@
-# Generated files 
-# (temporarily)
-dist/
