--- punchcard/ui/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-  <title>Testing React</title>
-</head>
-
-<body>
-  <h1>Hello!</h1>
-  <div id="root"></div>
-  <script src="src/app.tsx"></script>
-</body>
-
-</html>
