--- punchcard/ui/src/@types/react-heatmap-grid/index.d.ts
@@ -1,7 +1,0 @@
-declare module 'react-heatmap-grid' {
-  export default function HeatMap(
-      xLabels: Array<string>, 
-      yLabels: Array<string>, 
-      data: Array<number>): React.Component<any>;
-}
-
