--- punchcard/ui/src/App.test.tsx
@@ -1,14 +1,0 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-
-import 'jest-localstorage-mock';
-
-import App from './App';
-
-it('renders without crashing', () => {
-  sessionStorage.setItem('results', "{}")
-
-  const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
-  ReactDOM.unmountComponentAtNode(div);
-});
