--- punchcard/ui/webpack.config.js
@@ -1,1 +1,0 @@
-module.exports = { };
