--- tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "jsx": "react",
-    "lib": [
-      "es2015.promise",
-      "dom",
-      "es5"
-    ],
-    "rootDirs": [
-      ".",
-      "../bazel-bin/punchcard/ui"
-    ],
-    "types": []
-  }
-}
